flask_app/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
PyMySQL==1.1.1
Flask-SQLAlchemy==3.1.1
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
honeybadger @ git+https://github.com/honeybadger-io/honeybadger-python.git@insights-instrumentation#egg=honeybadger
